Add Tileset + TileLayer

+++ StateParser.cpp
@@ -30,7 +30,7 @@ void StateParser::parseObjects(XMLElement * stateRoot, vector<Entity*>* entities
 	}
 	if (Player * pl = dynamic_cast<Player *>(ent))
 	{
-		float jump = atof(stateRoot->Attribute("jump"));
+		float jump = (float)atof(stateRoot->Attribute("jump"));
 		pl->SetJump(jump);
 	}
 	EntityParams * params = new EntityParams(id.c_str(), x, y, width, height, numFrames, row);
